requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sqlalchemy==2.0.0
mysqlclient==2.2.0
fastapi_sso==0.7.2
python-jose==3.3.0
pytest==7.4.2